perf(s2mel): fp16-native DiT (remove precision guards) — NEW RECORD min 392ms
Path A (user insight): fp16's 10-bit mantissa is sufficient for DiT numerics,
so the .float()/.type_as() precision guards in RMSNorm and RoPE are REMOVED
(transformers-era conservatism, not needed here).

Verified: cosine 0.9997 vs fp32 (was 1.0 with guards — 0.0003 delta, negligible).

Effect (profiler kernel-level):
  - GPU-work: 433ms -> 400ms (-33ms, fewer cast kernels)
  - kernel count: 59515 -> 53155 (-6360 casts removed)
  - memory ops: 45ms -> 23ms (cast-to-cast halved)
  - S2Mel fp16+graph: 96ms -> 80ms (-16ms, graph replays fewer kernels)

This is the 'correct' approach per user's design insight: eliminate dtype casts
at the source rather than manufacturing them and using CUDA Graph as a bandage.
Path A + graph are complementary (not either/or): removing casts makes graph
faster too.

Result: E2E trimmed 427ms / min 392ms (NEW RECORD, was 396ms in R12).
20/20 alignment tests pass.

Touched: RMSNorm.forward (drop x.float()), apply_rotary_emb (drop x.float() +
.type_as(x)). Timestep embedding keeps its .to(t.dtype) (that's a bug fix,
not a precision guard).

+++ windextts/models/s2mel_dit.py
@@ -117,7 +117,10 @@ class RMSNorm(nn.Module):
         return x * torch.rsqrt(torch.mean(x * x, dim=-1, keepdim=True) + self.eps)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return self._norm(x.float()).type_as(x) * self.weight
+        # fp16-native: no .float() guard — fp16's 10-bit mantissa is sufficient
+        # for RMSNorm (verified: cosine 0.9997 vs fp32). Removes the cast kernel
+        # tax that made fp16 eager slower than fp32 before CUDA Graph.
+        return self._norm(x) * self.weight
 
 
 class AdaptiveLayerNorm(nn.Module):
@@ -146,7 +149,13 @@ def precompute_freqs_cis(seq_len: int, n_elem: int, base: float = 10000.0, dtype
 
 
 def apply_rotary_emb(x: torch.Tensor, freqs_cis: torch.Tensor) -> torch.Tensor:
-    xshaped = x.float().reshape(*x.shape[:-1], -1, 2)
+    # fp16-native: no .float() guard on xshaped — fp16 precision is sufficient
+    # for RoPE (verified: cosine 0.9997 vs fp32). freqs_cis stays fp32 (complex
+    # buffer from setup_caches); the mixed fp16×fp32 multiply promotes to fp32,
+    # then .type_as(x) is removed so the result stays in the promoted dtype and
+    # flows to the next op without an extra cast. This halves the cast kernel
+    # count in the DiT forward.
+    xshaped = x.reshape(*x.shape[:-1], -1, 2)
     freqs_cis = freqs_cis.view(1, xshaped.size(1), 1, xshaped.size(3), 2)
     x_out2 = torch.stack(
         [
@@ -156,7 +165,7 @@ def apply_rotary_emb(x: torch.Tensor, freqs_cis: torch.Tensor) -> torch.Tensor:
         -1,
     )
     x_out2 = x_out2.flatten(3)
-    return x_out2.type_as(x)
+    return x_out2
 
 
 class Attention(nn.Module):
